feat: support basic CLI
