--- examples/mysql/go.mod
@@ -1,10 +1,0 @@
-module github.com/api7/etcd-adapter/examples/mysql
-
-go 1.16
-
-require (
-	github.com/api7/etcd-adapter v0.1.1
-	go.uber.org/zap v1.19.1
-)
-
-replace github.com/api7/etcd-adapter v0.1.1 => ../../
