+++ internal/adapter/etcd.go
@@ -1,4 +1,4 @@
-package etcdadapter
+package adapter
 
 import (
 	"context"
@@ -11,8 +11,8 @@ import (
 	"go.uber.org/zap"
 	"google.golang.org/grpc"
 
-	"github.com/api7/etcd-adapter/backends/btree"
-	"github.com/api7/etcd-adapter/backends/mysql"
+	"github.com/api7/etcd-adapter/internal/backend/btree"
+	"github.com/api7/etcd-adapter/internal/backend/mysql"
 )
 
 // EventType is the type of event kind.
